ADD code for training

+++ loaders/kinetics_loader.py
@@ -92,7 +92,7 @@ class Kinetics(VisionDataset):
             return self.cached_graphs[idx]
 
         video, audio, info, video_idx = self.video_clips.get_clip(idx)
-        video = video * 1.0
+        video = (video * 1.0).to(torch.float32)
         label = self.samples[video_idx][1]
 
         if self.transform is not None:
